restAPI/requirements.txt
firebase-admin==6.0.1
Flask==2.2.2
Flask-Cors==3.0.10
google.cloud==0.34.0
pick==2.2.0
pylint==2.15.9
rich==12.6.0
